ENH: Streamline execution of tests downloading Slicer
This commit revisits the approach for running the self-tests. Instead of
attempting to build Slicer, the extension and then execute the tests, it
downloads the latest stable Slicer release, and then run the "TutorialMaker"
module tests specifying the `--additional-module-paths` parameter along
with the `--python-code` argument to directly use `slicer.testing.runUnitTest`.

Co-authored-by: Douglas Samuel Gonçalves <douglassam007@gmail.com>